coap-server/requirements.txt
CoAPthon3
requests
